+++ sbt-app/src/sbt-test/actions/input-task-dyn/build.sbt
@@ -12,7 +12,7 @@ lazy val root = (project in file(".")).
     name := "run-test",
     runFoo := Def.inputTaskDyn {
       val args = Def.spaceDelimited().parsed
-      (Compile / runMainBlock).toTask(s" Foo " + args.mkString(" "))
+      (Compile / runMain).toTask(s" Foo " + args.mkString(" "))
     }.evaluated,
     check := {
       val x = runFoo.toTask(" hi ho").value
