+++ sbt-app/src/sbt-test/dependency-management/exclude-dependencies2/test
@@ -1,3 +1,3 @@
 $ delete output
-> runBlock
+> run
 $ exists output
