+++ sbt-app/src/sbt-test/dependency-management/profiles/test
@@ -1,3 +1,3 @@
 $ delete output
-> runBlock
+> run
 $ exists output
