+++ sbt-app/src/sbt-test/lm-coursier/classifiers/test
@@ -1,1 +1,1 @@
-> runBlock
+> run
