+++ sbt-app/src/sbt-test/lm-coursier/from-no-head/test
@@ -1,3 +1,3 @@
 $ delete output
-> runBlock
+> run
 $ exists output
