+++ sbt-app/src/sbt-test/lm-coursier/from-wrong-url/test
@@ -1,3 +1,3 @@
 $ delete output
-> runBlock
+> run
 $ exists output
