+++ sbt-app/src/sbt-test/lm-coursier/from/test
@@ -1,4 +1,4 @@
 $ delete output
-> runBlock
+> run
 $ exists output
 $ delete shapeless_2.11-2.3.0.jar
