+++ sbt-app/src/sbt-test/lm-coursier/per-config-resolution/test
@@ -1,2 +1,2 @@
-> runBlock
+> run
 > Test/run
