+++ sbt-app/src/sbt-test/lm-coursier/scala-jars/test
@@ -1,3 +1,3 @@
 $ delete output
-> runBlock
+> run
 $ exists output
