+++ sbt-app/src/sbt-test/lm-coursier/scala-sources-javadoc-jars/test
@@ -1,4 +1,4 @@
 $ delete output
-> runBlock
+> run
 $ exists output
 > updateClassifiersCheck
