+++ sbt-app/src/sbt-test/lm-coursier/simple/test
@@ -1,4 +1,4 @@
 $ delete output
-> runBlock
+> run
 $ exists output
 > updateSbtClassifiers
