+++ sbt-app/src/sbt-test/lm-coursier/zookeeper/test
@@ -1,3 +1,3 @@
 $ delete output
-> runBlock
+> run
 $ exists output
