+++ sbt-app/src/sbt-test/source-dependencies/constants/test
@@ -3,14 +3,14 @@
 $ copy-file changes/B.scala B.scala
 
 $ copy-file changes/A1.scala A.scala
-> runBlock 1
+> run 1
 $ copy-file changes/A2.scala A.scala
-> runBlock 2
+> run 2
 
 > clean
 > ++2.13.12!
 
 $ copy-file changes/A1.scala A.scala
-> runBlock 1
+> run 1
 $ copy-file changes/A2.scala A.scala
-> runBlock 2
+> run 2
