+++ sbt-app/src/sbt-test/source-dependencies/implicit-search/test
@@ -2,10 +2,10 @@ $ copy-file changes/A1.scala A.scala
 $ copy-file changes/B.scala B.scala
 $ copy-file changes/C.scala C.scala
 > compile
--> runBlock
+-> run
 
 $ copy-file changes/A2.scala A.scala
 $ sleep 1000
 
 > compile
-> runBlock
+> run
