+++ sbt-app/src/sbt-test/source-dependencies/linearization/test
@@ -1,7 +1,7 @@
 > compile
 
 # the value of F.x should be 16
-> runBlock 16
+> run 16
 
 # modify D.scala so that the linearization changes
 $ copy-file changes/D.scala D.scala
@@ -12,4 +12,4 @@ $ sleep 1000
 
 # if F is recompiled, the value of x should be 11, otherwise it will still be 16
 # and this will fail
-> runBlock 11
+> run 11
