+++ sbt-app/src/sbt-test/source-dependencies/specialized/test
@@ -2,7 +2,7 @@
 > compile
 
 # verify that erased A.x can be called normally and reflectively
-> runBlock false
+> run false
 
 # make A.x specialized
 $ copy-file changes/A.scala A.scala
@@ -12,4 +12,4 @@ $ copy-file changes/A.scala A.scala
 
 # verify that specialized A.x can be called normally and reflectively
 # NOTE: this test doesn't actually work correctly: have to check the output to see that B.scala was recompiled
-> runBlock true
+> run true
